// projects/codx-ad/src/lib/models/AD_UserGroups.models.ts
export class AD_UserGroups{
    userID: string;
    userType: string;
    userName: string;
    userGroup: string;
    email: string;
    buid: string;
    activeOn: Date;
    phone: string;
    employeeID: string;
    stop: boolean;
    status: string;
    groupName: string;
    resource: string;
    buName: string;
    createdOn: Date;
    modifiedOn: Date;
    sessionID: string;
    owner: String;
    positionID: string;
    description: string;
}